Update and rename index.js to entry index.js
